wiktac-node-agent/agent/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
PyYAML==6.0.2
httpx==0.27.2
